Routines refactored.
Introduced two new abstact classes:
   AbstractCallbackList: as apposed to AbstractRoutine which takes a single callback in constructor AbstractCallbackList receives a keyed array of callbacks in the constructor and exposes them.
   AbstractCallbackMediator: Extends AbstractCallbackList and facilitates the choosing of an appropriate callback from the list. It mediates a "when" request authorzation by requiring 4 methods implemented by sub classes to perform the authorization these are:
           identifyRequested: Returns an array, empty or with a list of requested values
           considerProvisions: For each of the requested values the subclass my consider a list of possible provisions based on the requested entry.
           authorize: (not absract - can be overwritten) gets called for each of the provided values to compare if requested can be authorized based on provided (previausly called compareItems)
           notifyApproved: if authorized this method allows the subclasses to react on an approved state
           notifyDeclined: alternatively if authorization was declined allows subclasses to react on the new state (ie. set error headers etc.)

Refactored AbstractAccept to extend AbstractAccept class to extend AbstractCallbackMediator which reduced the complexity of the implementation considerably.
The Accept subclasses were reduced to only the functionality the specificly differ with, this boils down to the way the compare requested and provided in the authorize implementation because these vaalues may differ.
ContentType now does not have to extend AbstractAccept as it did before which was wrong in so many ways, it now extends AbstractCallbackMediator. This slightly increased the amount of code needed as it now cannot rely on Abstract accept anymore but these merely boils down to the new contractual methods that needs to be implemented and the change is not considerable but it is much more structured modular and understandable.
UserAgent also eroneously extended AbstractAccept and can now enjoy the functionality provided by the Callback Mediator without the overhead of the content negatiation.

Some additonal tests were added to test user agent and although more tests are required everything that we had works 100%