100% testcoverage but don't be fooled.
The last straw that breaks the donkey's back.
There is serious trouble wiht Content-Type because it get's implemented with Accept. It now has a bunch of bogus headers and still has a valid response even thaugh it shauld fail. Content Type also bad due to same issues. This is fixed in new branch see route-inspector.