appendRoutine should play nice with others.
The routines will all chain back the Route object to allow for more routines to be uhm chained along. All but appendRoutine were playing along. Instead of just addressing the issue, appendRoutine is liable for all the routines being uhm appended so now it became its responsibility to enable the chaining as a whole, taking some strain off AbstractRoutine in the process having less to be concerned about... and the pandas are happy again.