Better extension detection.
Previously we would only consider words ie .html or .json but .json-patch would miss.
According to [RFC3986](http://pretty-rfc.herokuapp.com/RFC3986#unreserved) the following are allowed:
ALPHA / DIGIT / - / . / _ / ~
These have been added as well as an additional symbol which is not ilegal perse as it is used in uri's.
The + is used in the schema and for denoting spaces in urls so it is perfectly legal to have an extension
of .hal+json
Since this is the main reason we are using this functionality as to map extensions to media types. As an alternative
because we are not looking up extensions and have no means to map these, this will allow us to use the whole subtype
as in type-name / subtype-name which will simplify doing the relation.
Either way it should be possible to map extonsions for language info.en.html info.en-US.html which we may now do.