Make routines accept instances that are callable.
In other words, instances of classes that implement the __invoke()
method.
To accept class names instead of only instances we need to change
the validation of callable objects in Routine\AbstractRoutine,
something that would be nice to have. :)

Also created a *src* directory inside tests to put classes related
to tests only (stubs). Let me know if this bothers you a lot, I
still think that this is better than declaring multiple namespaces
on files.