Fixes bug where OPTIONS req. calls other handlers
This commit fixes a bug where OPTIONS requests would, in addtion
to set the correct 'Allow' header, call a route that matches on
path, but not on method. More specifically the last handler that
matches on the path.

The commit also makes sure that custom OPTIONS handlers are called.
So if an application now registers an '->options' handler the correct
'Allow' header is set then the custom OPTIONS handler is called.